Allow to monitor all changes in repositories of a user (#12)
* Allow to monitor all changes in repositories of a user

Dear guys, thank you very much for your helpful Github integration. However, I would propose a small function extension:

# Current status
If I want to monitor the repositories of an user, I have to define each user repository as a single line in the config.py file. If one has lots of repositories this is quite time-consuming.

# Updated version
I can still set the behaviour for each user repository individually, but if the user repository has no individual configuration. A global configuration for all repositories of an user (if defined) is used. 
Furthermore, I replaced the if statements with  elif statements which is IMHO a little bit more the python way.